cleanup. add isHardArchived util
