+++ src/components/ProjectCard.tsx
@@ -1,15 +1,14 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import * as constants from '@ethersproject/constants'
 import { Skeleton } from 'antd'
-import { PV_V1, PV_V2 } from 'constants/pv'
-import { V1ArchivedProjectIds } from 'constants/v1/archivedProjects'
-import { V2ArchivedProjectIds } from 'constants/v2v3/archivedProjects'
+import { PV_V2 } from 'constants/pv'
 import { useProjectHandleText } from 'hooks/useProjectHandleText'
 import { useProjectMetadata } from 'hooks/useProjectMetadata'
 import useSubgraphQuery from 'hooks/useSubgraphQuery'
 import { ProjectTagName } from 'models/project-tags'
 import { Project } from 'models/subgraph-entities/vX/project'
 import Link from 'next/link'
+import { isHardArchived } from 'utils/archived'
 import { formatDate } from 'utils/format/formatDate'
 import { v2v3ProjectRoute } from 'utils/routes'
 
@@ -105,11 +104,7 @@ export default function ProjectCard({
       : projectCardHref
 
   const isArchived =
-    (projectCardData.pv === PV_V1 &&
-      V1ArchivedProjectIds.includes(projectCardData.projectId)) ||
-    (projectCardData.pv === PV_V2 &&
-      V2ArchivedProjectIds.includes(projectCardData.projectId)) ||
-    metadata?.archived
+    isHardArchived({ ...projectCardData }) || metadata?.archived
 
   const tags = (project as ProjectCardProject).tags
