+++ src/models/walletContribution.ts
@@ -1,0 +1,9 @@
+import { BigNumber } from '@ethersproject/bignumber'
+import { WalletContributionsQuery } from 'generated/graphql'
+
+import { PV } from './pv'
+
+export type WalletContribution = WalletContributionsQuery['participants'][0] & {
+  volume: BigNumber
+  pv: PV
+}
